+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 - FIXED #74 by adding support for UTF-8 in options #302
-- FIXED #305 by String interop conversions are now made in UTF-8 and avoid memory leaks #306
+- FIXED #305 by String interop conversions are now made in UTF-8 and avoid memory leaks #306 #310
 - FIXED #276 by changing NuGet package folder structure #309
 
 # 2.1.154